requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
dnspython==2.7.0
pillow==11.1.0
psycopg2-binary==2.9.10
pymongo==4.11.3
redis==5.2.1
sqlparse==0.5.3
